requirements.txt
streamlit
langchain
openai
python-dotenv
pandas
faiss-cpu
PyPDF2
docx
unstructured
tiktoken
langchain-openai
langchain-community
PyYAML
sqlite3
chromadb